Removing Static Library Dependency, removing lots of STL dependencies, Switching to mostly UE4 Native Memory Allocations
This change is meant to make the plugin remove the need for the compiled
static library. It should ease debugging as well as be more compatible
with UE4's native memory allocation schemes which should make it run
better on specific console hardware like PS4s with a dual GPU/CPU memory
bus architecture.